requirements.txt
streamlit
langchain-openai
langchain-huggingface
langchain-qdrant
qdrant-client
sentence-transformers
pillow
python-dotenv